Move client options logic into src/Config.
